Adds PrimeFactors kata and test
Done, but not sure how much I learned. Not much of a math guy so I just followed
along.  I didn't have to work out any of the kata myself, which sucks.  It was
nice to read and follow along with someone else's TDD, so it wasn't a total
loss.